refactor: :recycle: move Encode tests to itw own tester
